requirements.txt
openpyxl
pymongo
flask
uwsgi
requests
boto
flask-login
flask-sqlalchemy
